feat: cache-control calculation from directives
This change adds support for the @cacheControl directive in the form of a graphql-java instrumentation class, matching the behavior of the apollo-server's built-in cache control plugin:

https://github.com/apollographql/apollo-server/tree/291c17e255122d4733b23177500188d68fac55ce/packages/apollo-server-core/src/plugin/cacheControl

As well as the special _entities behavior in @apollo/subgraph:

https://github.com/apollographql/federation/blob/00c89f82bfcc2e27e6714d894874422f84ae468c/subgraph-js/src/types.ts#L76-L87

This change does not support setting cache hints dynamically in resolvers, only static cache hints in SDL directives.

The instrumentation emits the value in the response's extensions object. Apollo Gateway looks for a cache-control policy in a subgraph's HTTP headers, so it's up to the subgraph service owner to pluck the value off the extensions and add it as a response header.